all: fix tests for 32-bit architectures
Fix the tests to work on GOARCH=386.
Also add another test suite that tests on GOARCH=386.

Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/241537

+++ internal/encoding/text/decode_test.go
@@ -1021,10 +1021,10 @@ func TestDecoder(t *testing.T) {
 			in: `nums: [` +
 				fmt.Sprintf("%d", uint64(math.MaxUint64)) + `,` +
 				fmt.Sprintf("%d", uint32(math.MaxUint32)) + `,` +
-				fmt.Sprintf("%d", math.MaxInt64) + `,` +
-				fmt.Sprintf("%d", math.MinInt64) + `,` +
-				fmt.Sprintf("%d", math.MaxInt32) + `,` +
-				fmt.Sprintf("%d", math.MinInt32) +
+				fmt.Sprintf("%d", int64(math.MaxInt64)) + `,` +
+				fmt.Sprintf("%d", int64(math.MinInt64)) + `,` +
+				fmt.Sprintf("%d", int32(math.MaxInt32)) + `,` +
+				fmt.Sprintf("%d", int32(math.MinInt32)) +
 				`]`,
 			want: []R{
 				{K: text.Name},
